+++ .eslintrc.js
@@ -3,5 +3,8 @@ module.exports = {
   extends: ['standard-with-typescript'],
   parserOptions: {
     project: './tsconfig.json'
-  }
+  },
+  ignorePatterns: [
+    'dist'
+  ]
 }