+++ packages/vite-plugin-shopify/src/config.ts
@@ -1,14 +1,24 @@
 import path from 'path'
-import { Plugin, UserConfig } from 'vite'
+import { Plugin, UserConfig, mergeConfig } from 'vite'
 import glob from 'fast-glob'
+import createDebugger from 'debug'
 
 import { ResolvedVitePluginShopifyOptions } from './options'
+
+const debug = createDebugger('vite-plugin-shopify:config')
 
 // Plugin for setting necessary Vite config to support Shopify plugin functionality
 export default function shopifyConfig (options: ResolvedVitePluginShopifyOptions): Plugin {
   return {
     name: 'vite-plugin-shopify-config',
-    config () {
+    config (config: UserConfig): UserConfig {
+      const host = config.server?.host ?? 'localhost'
+      const port = config.server?.port ?? 5173
+      const https = config?.server?.https
+      const protocol = https === true ? 'https:' : 'http:'
+      const origin = `${protocol}//${host as string}:${port}`
+      const socketProtocol = https === true ? 'wss' : 'ws'
+
       const generatedConfig: UserConfig = {
         // Use relative base path so to load imported assets from Shopify CDN
         base: './',
@@ -19,8 +29,6 @@ export default function shopifyConfig (options: ResolvedVitePluginShopifyOptions
           outDir: path.join(options.themeRoot, 'assets'),
           // Do not use subfolder for static assets
           assetsDir: '',
-          // Clear output directory before each build
-          emptyOutDir: true,
           // Configure bundle entry points
           rollupOptions: {
             input: glob.sync(path.join(options.entrypointsDir, '**/*'), { onlyFiles: true })
@@ -31,16 +39,27 @@ export default function shopifyConfig (options: ResolvedVitePluginShopifyOptions
         resolve: {
           // Provide import alias to source code dir for convenience
           alias: {
-            '~': options.sourceCodeDir,
-            '@': options.sourceCodeDir
+            '~': path.resolve(options.sourceCodeDir),
+            '@': path.resolve(options.sourceCodeDir)
           }
         },
         server: {
-          host: true
+          host,
+          https,
+          port,
+          origin,
+          strictPort: true,
+          hmr: {
+            host: host as string,
+            port,
+            protocol: socketProtocol
+          }
         }
       }
 
-      return generatedConfig
+      debug(generatedConfig)
+
+      return mergeConfig(generatedConfig, config)
     }
   }
 }