+++ themes/seed-theme/frontend/entrypoints/theme.js
@@ -3,13 +3,17 @@
  */
 
 /* Modules */
+import '../../modules/announcement-bar/announcement-bar'
 import '../../modules/background-video/background-video'
 import '../../modules/cart-item/cart-item'
 import '../../modules/cart/cart'
-import '../../modules/collection-filters/collection-filters'
 import '../../modules/collection-product-grid/collection-product-grid'
 import '../../modules/hero-slider/hero-slider'
 import '../../modules/product-form/product-form'
 import '../../modules/product-gallery/product-gallery'
 import '../../modules/product-grid-toggle/product-grid-toggle'
+import '../../modules/product-main/product-main'
 import '../../modules/quantity-input/quantity-input'
+import '../../modules/recently-viewed-products/recently-viewed-products'
+import '../../modules/recommended-products/recommended-products'
+import '../../modules/site-header/site-header'