+++ themes/seed-theme/modules/collection-product-grid/collection-product-grid.js
@@ -1,4 +1,6 @@
-class CollectionProductGrid extends HTMLElement {
+import DynamicSectionElement from '@/scripts/dynamic-section-element.js'
+
+class CollectionProductGrid extends DynamicSectionElement {
   constructor () {
     super()
 
@@ -6,12 +8,11 @@ class CollectionProductGrid extends HTMLElement {
       paginationLoading: false
     }
 
-    this.addEventListener('click', this.onClickPaginationLink.bind(this))
-    this.addEventListener('click', this.onClickFilterLink.bind(this))
     this.addEventListener('change', this.onChangeFilter.bind(this))
+    this.addEventListener('click', this.onClickFilterLink.bind(this))
+    this.addEventListener('click', this.onClickPaginationLink.bind(this))
 
     this.loadMoreObserver = new IntersectionObserver(this.onLoadMoreScroll.bind(this))
-
     this.loadMoreObserver.observe(this.querySelector('[data-pagination]'))
   }
 
@@ -26,18 +27,20 @@ class CollectionProductGrid extends HTMLElement {
       event.target.closest('[data-pagination]').querySelector('[data-pagination-loading]').classList.remove('hidden')
       this.state.paginationLoading = true
 
-      fetch(`${event.target.href}&sections=collection-product-grid`)
-        .then((response) => this.update(response, true))
+      this.loadSectionFromUrl(event.target.href, {
+        appendToSlots: ['product-grid-items']
+      }).then(() => {
+        this.state.paginationLoading = false
+      })
     }
   }
 
-  // Handle pagination link click
+  // Handle filter link click
   onClickFilterLink (event) {
-    if (event.target.tagName === 'A' && event.target.closest('collection-filters')) {
+    if (event.target.tagName === 'A' && event.target.closest('[data-module="collection-filters"]')) {
       event.preventDefault()
 
-      fetch(`${event.target.href}&sections=collection-product-grid`)
-        .then(this.update.bind(this))
+      this.loadSectionFromUrl(event.target.href)
     }
   }
 
@@ -47,6 +50,18 @@ class CollectionProductGrid extends HTMLElement {
     if (event.target.closest('[data-module="collection-filters"]')) {
       // Generate new URL with filter query params
       const { origin, pathname } = window.location
+
+      let path = pathname
+      const tagFilters = []
+
+      if (path[path.length - 1] === '/') {
+        path = path.slice(0, path.length - 1)
+      }
+
+      if (pathname.includes('/filter_')) {
+        path = path.slice(0, path.indexOf('/filter_'))
+      }
+
       const params = new URLSearchParams(new FormData(event.target.closest('form')))
 
       for (const param of params) {
@@ -55,16 +70,24 @@ class CollectionProductGrid extends HTMLElement {
         if (!value) {
           params.delete(name)
         }
+
+        if (name === 'tag-filter') {
+          tagFilters.push(value)
+        }
       }
 
-      const newUrl = `${origin}${pathname}?${params}`
+      params.delete('tag-filter')
+
+      let fullPath = path
+
+      if (tagFilters.length) {
+        fullPath += `/${tagFilters.join('+')}`
+      }
+
+      console.log(`${origin}${fullPath}?${params}`)
 
       // Fetch new collection product grid section and update contents
-      fetch(`${newUrl}&sections=collection-product-grid`)
-        .then(this.update.bind(this))
-        .then(() => {
-          window.history.replaceState(null, null, newUrl)
-        })
+      this.loadSectionFromUrl(`${origin}${fullPath}?${params}`, { replaceState: true })
     }
   }
 
@@ -74,40 +97,9 @@ class CollectionProductGrid extends HTMLElement {
         const paginationlink = this.querySelector('[data-pagination] a')
 
         if (paginationlink) {
-          this.querySelector('[data-pagination] a').click()
+          paginationlink.click()
         }
       }
-    })
-  }
-
-  // Update section with newly-rendered content
-  update (response, appendProducts = false) {
-    response.json().then((sections) => {
-      const { 'collection-product-grid': newSectionHtml } = sections
-      const newSection = new DOMParser().parseFromString(newSectionHtml, 'text/html')
-
-      // Update product grid
-      if (appendProducts) {
-        // Append newly-loaded products to existing product grid
-        this.querySelector('[data-module="product-grid"] > div').insertAdjacentHTML(
-          'beforeend',
-          newSection.querySelector('[data-module="product-grid"] > div').innerHTML
-        )
-      } else {
-        // Replace product grid with new contents
-        this.querySelector('[data-module="product-grid"] > div').innerHTML =
-          newSection.querySelector('[data-module="product-grid"] > div').innerHTML
-      }
-
-      // Update collection filters
-      this.querySelector('collection-filters')
-        .update(newSection.querySelector('collection-filters'))
-
-      // Update pagination section
-      this.querySelector('[data-pagination]').innerHTML =
-        newSection.querySelector('[data-pagination]').innerHTML
-
-      this.state.paginationLoading = false
     })
   }
 }