+++ themes/seed-theme/modules/product-gallery/product-gallery.js
@@ -7,17 +7,18 @@ class ProductGallery extends HTMLElement {
       this.setActiveThumbnail()
     })
     this.addEventListener('focusin', (event) => {
-      console.log(event)
       this.setFocusedThumbnail()
     })
     this.addEventListener('focusout', (event) => {
-      console.log(event)
       this.setFocusedThumbnail()
     })
 
     // Set initial active thumbnail
-    this.setActiveThumbnail()
-    this.setFocusedThumbnail()
+
+    if (this.querySelectorAll('label[for^="selected-media"]').length > 1) {
+      this.setActiveThumbnail()
+      this.setFocusedThumbnail()
+    }
   }
 
   // Toggle active and inactive thumbnail styling based on current slide