+++ themes/seed-theme/modules/site-header/site-header.js
@@ -1,0 +1,4 @@
+import DynamicSectionElement from '@/scripts/dynamic-section-element.js'
+
+// Use base DynamicSectionElement class to enable replacing of content loaded from Section Rendering API
+customElements.define('site-header', DynamicSectionElement)