+++ themes/seed-theme/vite.config.js
@@ -6,13 +6,15 @@ import shopifyThemeSettings from 'vite-plugin-shopify-theme-settings'
 
 export default defineConfig({
   build: {
-    manifest: true
+    emptyOutDir: true,
+    sourcemap: true
   },
   plugins: [
-    // Add shopify plugin to enable entrypoint detection, smart generation of script and link tags, and support for Shopify CDN-hosted assets
+    // "Shopify" plugin enables entrypoint detection, smart generation of script and link tags, and support for Shopify CDN-hosted assets
     shopify(),
-    // Add shopify modules plugin to enable "modules" folder pattern collecting Liquid, CSS, and JS for theme modules
+    // "Shopify Modules" plugin enables "modules" folder pattern
     shopifyModules(),
+    // "Shopify Theme Settings" plugin enables managing theme settings schema in individual files under "config/src/*.json"
     shopifyThemeSettings()
   ]
 })