+++ themes/vite-shopify-example/assets/counter.d37853c9.js
@@ -1,0 +1,2 @@
+function c(n){let t=0;const e=o=>{t=o,n.innerHTML=`count is ${t}`};n.addEventListener("click",()=>e(++t)),e(0)}export{c as setupCounter};
+//# sourceMappingURL=counter.d37853c9.js.map