+++ themes/vite-shopify-example/frontend/components/counter.ts
@@ -1,0 +1,9 @@
+export function setupCounter (element: HTMLButtonElement): void {
+  let counter = 0
+  const setCounter = (count: number): void => {
+    counter = count
+    element.innerHTML = `count is ${counter}`
+  }
+  element.addEventListener('click', () => setCounter(++counter))
+  setCounter(0)
+}