+++ packages/vite-plugin-shopify-theme-settings/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vite-plugin-shopify-theme-settings
 
+## 0.0.2
+
+### Patch Changes
+
+- Update Vite dependency to 3.0.0 release
+
 ## 0.0.1
 
 ### Patch Changes