+++ packages/vite-shopify/CHANGELOG.md
@@ -1,5 +1,16 @@
 # vite-shopify
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - vite-plugin-shopify@0.0.2
+  - vite-plugin-shopify-modules@0.0.2
+  - vite-plugin-shopify-theme-settings@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes